requirements.txt
Django
django-registration
django-widget-tweaks
djangorestframework
django_excel
psycopg2
setuptools